requirements.txt
asyncio==3.4.3
python-dotenv==1.0.1
aiogram==3.13.1
aiohttp==3.9.5